requirements.txt
aiofiles==23.2.1
aiohttp==3.9.2
aiohttp-middlewares==2.2.1
aiohttp-socks==0.8.4
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.1.4
bleach==6.1.0
cachetools==4.2.4
cffi==1.16.0
click==8.1.7
defusedxml==0.7.1
dnspython==2.5.0
exceptiongroup==1.2.0
fastjsonschema==2.19.1
frozenlist==1.4.1
get-video-properties==0.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matrix-nio==0.24.0
mdurl==0.1.2
mistune==3.0.2
motor==3.3.2
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
opsdroid==0.28.0
opsdroid-get-image-size==0.2.2
packaging==23.2
pandocfilters==1.5.1
parse==1.20.1
peewee==3.17.0
platformdirs==4.1.0
puremagic==1.15
pycparser==2.21
pycron==3.0.0
pycryptodome==3.20.0
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-olm==3.2.16
python-socks==2.4.4
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tailer==0.4.1
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
unpaddedbase64==2.1.0
voluptuous==0.14.1
watchgod==0.8.2
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
zipp==3.17.0
